update: rename all id properties to id
