+++ src/main/java/com/huiun/fizzybudget/common/entity/UserRoleId.java
@@ -1,4 +1,4 @@
-package com.huiun.fizzybudget.common.entities;
+package com.huiun.fizzybudget.common.entity;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
