+++ src/main/java/com/huiun/fizzybudget/common/repository/RoleRepository.java
@@ -1,6 +1,6 @@
 package com.huiun.fizzybudget.common.repository;
 
-import com.huiun.fizzybudget.common.entities.Role;
+import com.huiun.fizzybudget.common.entity.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -8,8 +8,6 @@ import java.util.Optional;
 
 @Repository
 public interface RoleRepository extends JpaRepository<Role, Long> {
-
-    Optional<Role> findByRoleId(Long roleId);
 
     Optional<Role> findByRoleName(String roleName);
 }
