+++ src/main/java/com/huiun/fizzybudget/common/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package com.huiun.fizzybudget.common.repository;
 
-import com.huiun.fizzybudget.common.entities.User;
+import com.huiun.fizzybudget.common.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -8,8 +8,6 @@ import java.util.Optional;
 
 @Repository
 public interface UserRepository extends JpaRepository<User, Long> {
-
-    Optional<User> findByUserId(Long userId);
 
     Optional<User> findByUsername(String username);
