+++ src/main/java/com/huiun/fizzybudget/common/security/CustomUserDetails.java
@@ -1,6 +1,6 @@
 package com.huiun.fizzybudget.common.security;
 
-import com.huiun.fizzybudget.common.entities.User;
+import com.huiun.fizzybudget.common.entity.User;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
